Point calculation

+++ controllers/user_event.js
@@ -1,4 +1,15 @@
-const { User_Event, User, Event, Category, Admin } = require("../models");
+const {
+  Leaderboard,
+  User_Event,
+  User,
+  Event,
+  Category,
+  Admin,
+  Checkpoint,
+  AnswerQuiz,
+  sequelize,
+} = require("../models");
+const event = require("../routers/event");
 
 module.exports = class userEventController {
   static getAllEvents = async (req, res, next) => {
@@ -7,7 +18,7 @@ module.exports = class userEventController {
         include: [
           {
             model: User,
-            attributes: { exclude: ['password'] }
+            attributes: { exclude: ["password"] },
           },
           {
             model: Event,
@@ -17,11 +28,11 @@ module.exports = class userEventController {
               },
               {
                 model: Admin,
-                attributes: { exclude: ['password'] }
-              }
-            ]
-          }
-        ]
+                attributes: { exclude: ["password"] },
+              },
+            ],
+          },
+        ],
       });
       res.status(200).json(dataEvents);
     } catch (error) {
@@ -37,7 +48,7 @@ module.exports = class userEventController {
         include: [
           {
             model: User,
-            attributes: { exclude: ['password'] }
+            attributes: { exclude: ["password"] },
           },
           {
             model: Event,
@@ -46,32 +57,107 @@ module.exports = class userEventController {
                 model: Category,
               },
               {
+                model: Checkpoint, //! Tambahan !
+              },
+              {
                 model: Admin,
-                attributes: { exclude: ['password'] }
-              }
-            ]
-          }
-        ]
+                attributes: { exclude: ["password"] },
+              },
+            ],
+          },
+        ],
       });
-      res.status(200).json(dataEvent);
+
+      if (!dataEvent) throw ({ name: "Data not found!" })
+
+      const dataCheckpoint = await Checkpoint.findAll({
+        where: { EventId: id },
+      });
+      res.status(200).json({ dataEvent, dataCheckpoint });
     } catch (error) {
       console.log(error);
       next(error);
     }
   };
 
+  static getUserEventsByUserId = async (req, res, next) => {
+    try {
+      const { id } = req.user;
+      const eventsOfUser = await User_Event.findAll({
+        where: {
+          UserId: id,
+        },
+        include: [
+          {
+            model: Event,
+          },
+        ],
+      });
+      res.status(200).json(eventsOfUser);
+    } catch (error) {
+      next(error);
+    }
+  };
 
   static addEvent = async (req, res, next) => {
+    //!create 3 answer
+
+    const t = await sequelize.transaction();
     try {
       const { event_id } = req.params;
-      await User_Event.create({
-        UserId: req.user.id,
-        EventId: event_id
-      });
-      res.status(201).json({
-        message: `event successfully added`
-      });
+
+      const checkEvent = await User_Event.findOne(
+        {
+          where: {
+            UserId: req.user.id,
+            EventId: event_id,
+          },
+        },
+        { transaction: t }
+      );
+      if (!checkEvent) {
+        const checkpointsData = await Checkpoint.findAll({
+          where: { EventId: event_id },
+        });
+        const answerData = await checkpointsData.map((e) => {
+          return {
+            UserId: req.user.id,
+            CheckpointId: e.id,
+          };
+        });
+        await AnswerQuiz.bulkCreate(answerData, { transaction: t });
+        await User_Event.create(
+          {
+            UserId: req.user.id,
+            EventId: event_id,
+          },
+          { transaction: t }
+        );
+
+        const data = await Event.findByPk(event_id, { transaction: t });
+        const finalAddAmmount = data.amount + 100000;
+        await Event.update(
+          { amount: finalAddAmmount },
+          {
+            where: {
+              id: event_id,
+            },
+          },
+          { transaction: t }
+        );
+
+        await t.commit();
+        res.status(201).json({
+          message: `event successfully added`,
+        });
+      } else {
+        await t.commit();
+        res.status(403).json({
+          message: `event already added`,
+        });
+      }
     } catch (error) {
+      await t.rollback();
       console.log(error);
       next(error);
     }
@@ -80,8 +166,10 @@ module.exports = class userEventController {
   static deleteEvent = async (req, res, next) => {
     try {
       const { id } = req.params;
+      const user_event = User_Event.findByPk(id);
+      if (user_event) throw ({ name: "Data not found!" })
       await User_Event.destroy({
-        where: { id }
+        where: { id },
       });
       res.status(200).json({ message: "event successfully deleted" });
     } catch (error) {
@@ -90,4 +178,98 @@ module.exports = class userEventController {
     }
   };
 
+  // static getEventByUSersId = async (req, res, next) => {
+  //   try {
+  //     const dataEvents = await User_Event.findOne(
+  //       {
+  //         include: [
+  //           {
+  //             model: User,
+  //             attributes: { exclude: ["password"] },
+  //           },
+  //           {
+  //             model: Event,
+  //             include: [
+  //               {
+  //                 model: Category,
+  //               },
+  //               {
+  //                 model: Admin,
+  //                 attributes: { exclude: ["password"] },
+  //               },
+  //             ],
+  //           },
+  //         ],
+  //       },
+  //       { where: { UserId: req.user.id } }
+  //     );
+  //     const checkpointData = await Checkpoint.findAll({
+  //       where: { EventId: dataEvents.EventId },
+  //     });
+  //     const answerQuizData = await AnswerQuiz.findAll({
+  //       where: { UserId: req.user.id },
+  //     });
+  //     res.status(200).json({ dataEvents, checkpointData, answerQuizData });
+  //   } catch (error) {
+  //     console.log(error);
+  //     next(error);
+  //     //check
+  //   }
+  // };
+
+  static getEventByEventId = async (req, res, next) => {
+    try {
+      const findAnswerQuiz = async (checkpointIds) => {
+        return Promise.all(
+          checkpointIds.map(async (checkpointId) => {
+            return await AnswerQuiz.findOne({
+              where: { UserId: req.user.id, CheckpointId: checkpointId },
+            });
+          })
+        ).then((quizData) => {
+          return quizData;
+        });
+      };
+      const { id } = req.params;
+      const dataEvents = await User_Event.findOne(
+        {
+          include: [
+            {
+              model: User,
+              attributes: { exclude: ["password"] },
+            },
+            {
+              model: Event,
+              include: [
+                {
+                  model: Category,
+                },
+                {
+                  model: Admin,
+                  attributes: { exclude: ["password"] },
+                },
+              ],
+            },
+          ],
+        },
+        { where: { EventId: id } }
+      );
+
+      if (!dataEvents) throw ({ name: "Data not found!" })
+
+      const checkpointData = await Checkpoint.findAll({
+        where: { EventId: dataEvents.EventId },
+      });
+      const checkpointIds = checkpointData.map((el) => el.id);
+      const answerQuizData = await findAnswerQuiz(checkpointIds);
+      const leaderboard = await Leaderboard.findOne({ where: { EventId: id } });
+      res
+        .status(200)
+        .json({ dataEvents, checkpointData, answerQuizData, leaderboard });
+    } catch (error) {
+      console.log(error);
+      next(error);
+      //check
+    }
+  };
 };
